ci: add caching actions
Add "Restore" and "Save" caching actions.

These reduce boilerplate in the soon-to-be lengthy ci.yml configuration
file.

Using actions also allows us to update caching strategies in the
future in a single location.

These actions are implemented so that caches will be saved on pushes
to master only.

When a PR is opened it will cache hit on the caches from the lastest
push to master, or in the case of depends will hit on any matching
depends hash, falling back to partial matches in each case.

Depends caches are hashed ~ matching our own makefile rules there, i.e.
using a hash of the following files:

```
hashFiles('depends/packages/*.mk', 'depends/Makefile', 'depends/config.guess', 'depends/config.sub', 'depends/funcs.mk', 'depends/builders/*.mk', 'depends/hosts/*.mk')
```

This means we direct cache hit in cases where depends would not be
re-built.

The cirruslabs cache action will fallback transparently to GitHub's
cache in the case that the job is not being run on a Cirrus Runner,
making these compatible with running on forks (on free GH hardware).