ci: use buildx in ci
Despite buildkit being the default in docker for some number of releases
now, testing has found that unless the `docker buildx build` command
exactly is used, the incorrect docker builder will be selected.

Amend the ci script to use `docker buildx build` command when the env
var `$USE_BUILDX` is set, and set this in the CI workflow.