ci: add caching actions
Add specific "Restore" and "Save" caching actions.

These reduce boilerplate in the soon-to-be lengthy ci.yml configuration
file.

Using actions also allows us to updating caching strategies in the
future with a single change.

These actions are implemented so that caches will be saved on pushes
(i.e. *not* pull requests)l, and if we did not cache-hit on restore.

This means that in the main repo we will save ccache, depends-sources
and built-depends on pushes. Primarily this hits pushes to master
(merges). On forks pushes to branches will be saved too, enabling faster
re-runs of CI in your own forks.

When a PR is opened it will cache hit on the caches from the lastest
push to master, or in the case of depends will hit on any matching
depends hash.

Depends caches are hashed ~ matching our own makefile rules there, i.e.
using a hash of the following files:

```
hashFiles('depends/packages/*.mk', 'depends/Makefile', 'depends/config.guess', 'depends/config.sub', 'depends/funcs.mk', 'depends/builders/*.mk', 'depends/hosts/*.mk')
```

This means we cache hit in cases where depends would not be re-built.

The cirruslabs cache action will fallback transparently to GitHub's
cache in the case that the job is not being run on a Cirrus Runner,
making these compatible with running on forks (on free GH hardware).