ci: add caching actions
Add "Restore" and "Save" caching actions.

These actions reduce boilerplate in the main ci.yml configuration file.

These actions are implemented so that caches will be saved on `push`
only.

When a pull request is opened it will cache hit on the caches from the
lastest push (preferring master branch), or in the case of depends will
hit on any matching depends hash, falling back to partial matches.

Depends caches are hashed using `git rev-parse HEAD:depends`.
This means we direct cache hit in cases where depends would not be
re-built, otherwise falling back to a partial match.

The cirruslabs cache action will fallback transparently to GitHub's
cache in the case that the job is not being run on a Cirrus Runner,
making these compatible with running on forks (on free GH hardware).