doc: Improve doc for candidate blocks
It is not true in general that all blocks with the same work
(as defined via nChainWork) as the tip are candidates expected to be in
setBlockIndexCandidates - only ones with a better tiebreak than the tip.
Make that clear in various spots.

Also explain that we don't use CBlockIndexWorkComparator tiebreak rules
for m_best_header.