ci: add configure-docker action
Another action to reduce boilerplate in the main ci.yml file.

This action will set up a docker builder compatible with caching build
layers to a container registry using the `docker-container` build
driver.

The action will then determine whether it should log in to a container
registry or not.

Finally it will configure the build cache args.

When logged in build cache layers will be pulled and pushed to the
registry.

When not logged in build cached layers will only be pulled from the
registry.